+++ board/recalbox/fsoverlay/recalbox/scripts/recalbox-utils.sh
@@ -86,3 +86,14 @@ isRecalboxRGBDual() {
     return 1
   fi
 }
+
+# Returns true if the platform may need to force enable DRM connector.
+# This is related to black screen issue
+doesBoardNeedKMSEnabler() {
+  case "$(getArchName)" in
+    rpizero2|rpi3|rpi4|rpi4_64)
+      return 0 ;;
+    *)
+      return 1 ;;
+  esac
+}